The Elder Scrolls: Arena — Update script to 2.11 syntax

+++ play.it-2/games/play-the-elder-scrolls-1-arena.sh
@@ -1,8 +1,8 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
-# Copyright (c) 2015-2018, Antoine Le Gonidec
+# Copyright (c) 2015-2019, Antoine Le Gonidec
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -34,14 +34,12 @@ set -o errexit
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-script_version=20190302.1
+script_version=20190302.2
 
 # Set game-specific variables
 
 GAME_ID='the-elder-scrolls-1-arena'
 GAME_NAME='The Elder Scrolls: Arena'
-
-ARCHIVES_LIST='ARCHIVE_GOG'
 
 ARCHIVE_GOG='setup_tes_arena_2.0.0.5.exe'
 ARCHIVE_GOG_URL='https://www.gog.com/game/the_elder_scrolls_iii_morrowind_goty_edition'
@@ -49,20 +47,20 @@ ARCHIVE_GOG_MD5='ca5a894aa852f9dbb3ede787e51ec828'
 ARCHIVE_GOG_SIZE='130000'
 ARCHIVE_GOG_VERSION='1.07-gog2.0.0.5'
 
-ARCHIVE_DOC1_PATH='app'
-ARCHIVE_DOC1_FILES='./*.pdf ./readme.txt'
+ARCHIVE_DOC0_DATA_PATH='app'
+ARCHIVE_DOC0_DATA_FILES='*.pdf readme.txt'
 
-ARCHIVE_DOC2_PATH='tmp'
-ARCHIVE_DOC2_FILES='./gog_eula.txt'
+ARCHIVE_DOC1_DATA_PATH='tmp'
+ARCHIVE_DOC1_DATA_FILES='gog_eula.txt'
 
 ARCHIVE_GAME_BIN_PATH='app'
-ARCHIVE_GAME_BIN_FILES='./*.cfg ./*.exe ./*.inf ./*.ini'
+ARCHIVE_GAME_BIN_FILES='*.cfg *.exe *.inf *.ini'
 
-ARCHIVE_GAME_DATA1_PATH='app'
-ARCHIVE_GAME_DATA1_FILES='./*.65 ./*.ad ./*.adv ./*.bak ./*.bnk ./*.bsa ./*.cel ./*.cif ./*.clr ./*.col ./*.cpy ./*.dat ./*.flc ./*.gld ./*.ico ./*.img ./*.lgt ./*.lst ./*.me ./*.mif ./*.mnu ./*.ntz ./*.opl ./*.rci ./*.txt ./*.voc ./*.xfm ./cityintr ./citytxt ./extra ./speech'
+ARCHIVE_GAME0_DATA_PATH='app'
+ARCHIVE_GAME0_DATA_FILES='*.65 *.ad *.adv *.bak *.bnk *.bsa *.cel *.cif *.clr *.col *.cpy *.dat *.flc *.gld *.ico *.img *.lgt *.lst *.me *.mif *.mnu *.ntz *.opl *.rci *.txt *.voc *.xfm cityintr citytxt extra speech'
 
-ARCHIVE_GAME_DATA2_PATH='app/__support'
-ARCHIVE_GAME_DATA2_FILES='./save'
+ARCHIVE_GAME1_DATA_PATH='app/__support'
+ARCHIVE_GAME1_DATA_FILES='save'
 
 GAME_IMAGE='.'
 GAME_IMAGE_TYPE='cdrom'
@@ -74,7 +72,6 @@ APP_MAIN_EXE='acd.exe'
 APP_MAIN_OPTIONS='-Ssbpdig.adv -IOS220 -IRQS7 -DMAS1 -Mgenmidi.adv -IOM330 -IRQM2 -DMAM1'
 APP_MAIN_PRERUN='d:'
 APP_MAIN_ICON='goggame-1435828982.ico'
-APP_MAIN_ICON_RES='16 32 48 256'
 
 PACKAGES_LIST='PKG_DATA PKG_BIN'
 
@@ -84,51 +81,53 @@ PKG_DATA_DESCRIPTION='data'
 PKG_DATA_PROVIDE='arena-data'
 
 PKG_BIN_ARCH='32'
-PKG_BIN_DEPS_DEB="$PKG_DATA_ID, dosbox"
-PKG_BIN_DEPS_ARCH="$PKG_DATA_ID dosbox"
+PKG_BIN_DEPS="$PKG_DATA_ID dosbox"
 # Easier upgrade from packages generated with pre-20190302.1 scripts
 PKG_BIN_PROVIDE='arena'
 
 # Load common functions
 
-target_version='2.0'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
+prepare_package_layout
+rm --recursive "$PLAYIT_WORKDIR/gamedata"
 
-PKG='PKG_BIN'
-organize_data 'GAME_BIN' "$PATH_GAME"
+# Extract icons
 
 PKG='PKG_DATA'
-organize_data 'DOC1'       "$PATH_DOC"
-organize_data 'DOC2'       "$PATH_DOC"
-organize_data 'GAME_DATA1' "$PATH_GAME"
-organize_data 'GAME_DATA2' "$PATH_GAME"
-
-extract_and_sort_icons_from 'APP_MAIN'
-
-rm --recursive "$PLAYIT_WORKDIR/gamedata"
+icons_get_from_package 'APP_MAIN'
 
 # Write launchers
 
 PKG='PKG_BIN'
-write_launcher 'APP_MAIN'
+launcher_write 'APP_MAIN'
 
 # Build package
